Functional Options: Don't recommend closures (#77)
This changes the Functional Options example to not recommend closures,
instead favoring full implementations of the interface for each option.

To help demonstrate that options don't have to be new types, a logger
option was added to the example. This made the examples too long so the
the timeout option was dropped; cache was kept around to demonstrate
readability of naked booleans versus `WithCache(false)`.

Further, the function was renamed to Open and the Bad/Good side-by-side
comparison was reduced to just the user-facing API. The implementation
was moved to its own paragraph.